0.4.5
Response unmarshalling using JDK 1.6 XML stream, although only supported by LoadArtist and LoadAlbum so far. Have seen speeds 5x-20x faster than the previous DOM parser. The bottleneck now is retrieving and gunzipping data.

The unmarshalling is failfast, if an new tag is added to the response it needs to be handled in the code or it will throw an exception. This should be turned on and off using a setting in the future.

Of course this feature made that I found all this new data in the response. similar albums, reviews, external ids, and what not. Some of this data is skipped (especially links) but most of it is added to the domain objects. 

Some data is shared by many classes but not extended from a single class. Need to think that though I think RestrictedMedia should be renamed to something else and also include links, external ids, et c.

The XML schemas should be up to date with the data I've encountred, they do however need to be cleaned up.